Menu View and Game Creation/Joining
Added variables to Game that, for each of two players, stores an ID, a username, and a Grid. Game also has an inProgress variable. Added basic "New Game" and "Logout" buttons to the MenuPanel. When "New Game" pressed in MenuPanel, MenuControl informs GameServer. If a game is already in progress with two players, user gets error. Otherwise, user is moved to LobbyPanel and the GameServer stores user's information in a Game object.