Bug fixing; Bud identifying
The follow bug exists: If a Lobby/Game has exactly one user and that user leaves, the Lobby/Game does not update that they left. It still thinks the user is in the Lobby/Game, even though they have returned to the menu. Thus, when another player joins, the Lobby/Game will be full even though only the new user is in the lobby.