(POST-DUE-DATE): Two major bug fixes
Fixed a client-side bug in GameControl.endGame() where the GamePanel was not correctly reset after a game ended; this resulted in ships and hits/misses from the previous game persisting on-screen when a new game begins. Also fixed a server-side bug in  Game.removePlayer() that would not remove the player if playerTwo was null and playerOne was playerId. Previously, it would check if playerId matched playerTwoId first in the if-else statement. However, if playerTwoId was null, this.playerTwoId.equals(playerId)  threw a nullpointexception (that wasn't being shown anywhere for some reason), and none of the rest of the method would run. So, the player would be removed from the game client-side but the server would still think they were in a game. As a result, two players could not successfully join a lobby together after that point as the second player would see "Game already in progress."